added bikes to Map.tsx
