+++ src/models/pricingModel.tsx
@@ -6,6 +6,23 @@ const pricingModel = {
 
     return user.data;
   },
+  updatePricing: async function updatePricing(id: string, priceModel: any) {
+    fetch(`http://localhost:4000/pricing/${id}`, {
+      method: 'PUT',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(priceModel),
+    })
+      .then((res) => {
+        console.log(res);
+        res.json();
+      })
+      .catch((err) => {
+        return err;
+      });
+    return 'success';
+  },
 };
 
 export default pricingModel;
